Several QOL changes
Added argument parsing to allow for better general use.

Lowered threshold to accommodate the new model.

used round() function in "Unknown" label return condition.

Added a condition to the main loop, for it to close when the streaming window is closed.

Changed the required number of frames to pass for the person to be changed.

added an option to be used with images, so that one can run the model on an image and return the model's output.